MetatagManager to use entity URL when possible
Currently the MetatagManager uses the current language returned from the language manager. This is a problem when one wants to programmatically generate metatags for multiple pages in multiple languages that use tokens (The tokens are all in the same language and not in the language of the content).

We also discovered this issue when using the GraphQL module. It fails to properly generate the properly translated metatags because it doesn't switch the "current language" as it goes through the various nodes and their translations.

This fix uses the entity language (if one exists) when doing the token replacement, otherwise it will use the default current language that it is using now.